+++ include/common.hpp
@@ -1,6 +1,10 @@
 #pragma once
 
+#include "util/version.hpp"
+
 namespace common {
+
+inline constexpr Version VERSION{0, 3, 2};
 
 inline bool is_running{true};